Refactored add_transactions into load_transactions
