+++ src/fintrackr/fin_db.py
@@ -240,96 +240,6 @@ class FinDB:
                raise ValueError("Could not insert new data source in data_sources table")
         
         return source_name_tuple[0][0]
-
-    def add_transactions(self, path_to_source_file: str, source_info: str) -> None:
-        """
-        Load transactions from a file and log the addition of these transactions
-        in the data_load_metadata table.
-
-        Only new transactions are added; duplicates (which have identity across the 3
-        input columns of Date, Amount, and Description with an existing transaction row)
-        are ignored.
-
-        Parameters
-        ----------
-        path_to_source_file: str
-            Path to a csv where every row is a transaction.
-        source_info: str
-            Are these transactions from credit card, checking account, etc
-            This is the "name" field in the data_sources table.
-            It will be added if it doesn't already exist.
-
-        Returns
-        -------
-        int
-            Number of transactions added.
-        """
-
-        num_new_transactions = 0
-
-        staging_cols = [Col_Def(col_name="posted_date", col_type="date"),
-                Col_Def(col_name="amount", col_type="money"),
-                Col_Def(col_name="description", col_type="text")
-        ]
-
-        num_staged_transactions = self.csv_to_staging(csv_path=path_to_source_file, csv_columns=staging_cols)
-
-        if num_staged_transactions == 0:
-            logger.info("No transactions loaded from source file to staging table; no transactions will be added")
-            return num_new_transactions
-        
-        # Get id for this source_info or add if it doesn't exist
-        source_info_id = self.add_data_source(source_info)
-
-        today_date = date.today()
-        
-        transactions_query = "WITH joined AS ( " \
-            "    SELECT s.* " \
-            "    FROM staging s " \
-            "    LEFT JOIN transactions t ON " \
-            "        t.posted_date = s.posted_date AND " \
-            "        t.amount = s.amount AND " \
-            "        t.description = s.description " \
-            "    WHERE t.id IS NULL " \
-            "), " \
-            "meta AS ( " \
-            "    INSERT INTO data_load_metadata " \
-            "        (date_added, username, source, data_source_id) " \
-            "    VALUES (%s, %s, %s, %s)" \
-            "    " \
-            "    RETURNING id " \
-            ") " \
-            "INSERT INTO transactions (posted_date, amount, description, metadatum_id) " \
-            "SELECT posted_date, amount, description, meta.id " \
-            "FROM joined, meta " \
-            "RETURNING *;"
-            
-        try:
-            all_new_transactions = self.execute_query(transactions_query, (today_date, self.user, path_to_source_file, source_info_id))
-        except Exception as e:
-            logger.exception(f"Insertion into transactions table failed with exception: {e}; return from query: {num_new_transactions}")
-            raise ValueError(f"Insertion into transactions table failed with exception: {e}")
-        
-        if all_new_transactions is None:
-            logger.error("No transactions inserted")
-            # Check if all new transactions to load are already in db and that's why it failed:
-            check_dups = "SELECT s.* " \
-                "    FROM staging s " \
-                "    LEFT JOIN transactions t ON " \
-                "        t.posted_date = s.posted_date AND " \
-                "        t.amount = s.amount AND " \
-                "        t.description = s.description " \
-                "    WHERE t.id IS NULL;"
-            if len(self.execute_query(check_dups)) == 0:
-                logger.error("All staged transactions are already in transactions table")
-                return 0
-            else:
-                raise ValueError("No transactions inserted, but not because all new transactions were in db already")
-
-        # Drop staging table
-        self.execute_action("DROP TABLE staging;")
-
-        return len(all_new_transactions)
     
     def data_from_date_range(self, data_source: str, date_range: List[date]) -> dict[List[Transaction]]:
         """
