+++ tests/test_db.py
@@ -11,6 +11,7 @@ from datetime import date
 import fintrackr.testing_utils as utils
 from fintrackr.utils import Col_Def
 from fintrackr.load_balances import add_balance
+from fintrackr.load_transactions import add_transactions
 
 class TestDB(unittest.TestCase):
     @classmethod
@@ -73,40 +74,6 @@ class TestDB(unittest.TestCase):
     def test_add_data_source(self):
         # This is tested in several other places
         pass
-    
-    def test_add_transactions(self):
-        # Add_transactions calls csv_to_staging (which we test separately above)
-
-        num_transactions_added = self.FinDB.add_transactions(
-            path_to_source_file = self.path_to_test_transactions, 
-            source_info = self.source_info
-            )
-        self.assertEqual(num_transactions_added, self.transactions_to_add.shape[0], "Number of added transactions does not match file")
-        self.assertEqual(self.element_to_match, 
-                         self.FinDB.execute_query("SELECT amount FROM transactions WHERE description=%s;",('Concert tickets',))[0][0], 
-                         "Data were scrambled when loaded into transactions"
-                         )
-        
-        # Test that trying to upload the same file again fails
-        # (it actually errors out with a silent error, unfortunately: raises "Key (source)=(/Users/steph/Documents/Code/FinTrackr/tests/data/test_data_cc.csv) already exists)")
-        num_transactions_added = self.FinDB.add_transactions(
-            path_to_source_file = self.path_to_test_transactions, 
-            source_info = self.source_info
-            )
-        self.assertEqual(num_transactions_added, 0, "Duplicates should not have been successfully loaded")
-
-        # Test what happens when partial duplicates are added
-        additional_transactions_path = os.path.join(utils.TEST_DATA_PATH,"test_data_checking.csv")
-        addtl_trans = pd.read_csv(additional_transactions_path, header=None)
-        num_new_trans = len(addtl_trans)
-        dup_trans = self.transactions_to_add.loc[self.transactions_to_add.iloc[:,2]=="Safeway"]
-        addtl_trans = pd.concat([addtl_trans, dup_trans])
-
-        num_transactions_added = self.FinDB.add_transactions(
-            path_to_source_file = additional_transactions_path, 
-            source_info = self.source_info
-            )
-        self.assertEqual(num_transactions_added, num_new_trans, "Duplicates should not have been successfully loaded")
 
     def test_data_from_date_range(self):
         # pytest runs each test case independently, so re-set-up the db
@@ -118,7 +85,8 @@ class TestDB(unittest.TestCase):
             bal_date=self.balance_date,
             bal_amt=self.balance_amount
         )
-        self.FinDB.add_transactions(
+        add_transactions(
+            FinDB = self.FinDB,
             path_to_source_file = self.path_to_test_transactions, 
             source_info = self.source_info
         )
